fix bug

+++ app.py
@@ -144,7 +144,8 @@ def main():
     button_pressed_down = False
     pos = ""
     generate = 0 #generate image 0 is false, 3 is true
-    os.remove("result.png")
+    if os.path.exists("result.png"):
+        os.remove("result.png")
     while True:
         fps = cvFpsCalc.get()
